Rename diffusion kernel sources to diff_abs_nsteps
