+++ Diff_Abs_Nsteps/compile_commands.json
@@ -8,9 +8,9 @@
       "-hls-syn-headers-dir=C:\\Xilinx\\2025.1\\Vitis\\common\\technology\\autopilot",
       "-fhlstoplevel=diff_abs_nsteps",
       "-c",
-      "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_losses.cpp"
+      "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_abs_nsteps.cpp"
     ],
-    "file": "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_losses.cpp"
+    "file": "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_abs_nsteps.cpp"
   },
   {
     "directory": "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps",
@@ -20,8 +20,8 @@
       "-fhls-tb",
       "-hls-sim-headers-dir=C:\\Xilinx\\2025.1\\Vitis\\include",
       "-c",
-      "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_losses_tb.cpp"
+      "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_abs_nsteps_tb.cpp"
     ],
-    "file": "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_losses_tb.cpp"
+    "file": "C:\\Users\\Usuario\\Desktop\\DTI_work_space\\Diff_Abs_Nsteps\\.\\diff_abs_nsteps_tb.cpp"
   }
 ]
