+++ Diff_Abs_Nsteps/hls_config.cfg
@@ -7,14 +7,14 @@ package.output.syn=false
 clock=120Mhz
 clock_uncertainty=1%
 syn.top=diff_abs_nsteps
-syn.file=./diff_losses.cpp
-syn.file=./diff_losses.h
+syn.file=./diff_abs_nsteps.cpp
+syn.file=./diff_abs_nsteps.h
 csim.clean=1
 package.ip.vendor=fotonica109
 package.ip.version=1.0
 package.ip.description=operador de difraccion ADI
 package.ip.display_name=DiffAbsNsteps
 package.output.file=C:\\Vws25\\custom_ip_library\\diff_abs_nsteps
-tb.file=./diff_losses_tb.cpp
+tb.file=./diff_abs_nsteps_tb.cpp
 tb.file=./phi_in.dat
 tb.file=./golden.dat
